// client/src/types/Encounter.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Combatant } from "./Combatant";

export type Encounter = { id: string, campaign_id: string, active: boolean, current_turn_index: number, round_number: number, combatants: Array<Combatant>, created_at: string, };